refactor: added pages for routes with different base
