+++ client/src/components/ui/check-auth-link.tsx
@@ -1,6 +1,6 @@
 import { useAuthStore } from '@/store/authStore'
 import { useUiStore } from '@/store/useUiStore'
-import { Link, LinkProps } from '@tanstack/react-router'
+import { Link } from '@tanstack/react-router'
 
 export function CheckAuthLink({ children, className, ...props }: LinkProps) {
   const { isAuth } = useAuthStore()
