+++ client/src/components/ui/header/navigation.tsx
@@ -1,4 +1,4 @@
-import { routes } from '@/routes/routes'
+import { routes } from '@/router/routes'
 import { cn } from '@/lib/utils'
 import { Button } from '../shadcn-ui/button'
 import { Link } from '@tanstack/react-router'
@@ -25,7 +25,7 @@ export function Navigation({ className }: NavProps) {
           key={i}
           asChild
           variant="ghost"
-          className="h-[20px] transition-all duration-300 hover:font-bold hover:text-purple-400 max-lg:block max-lg:px-2 max-lg:py-1 max-lg:text-lg lg:text-base lg:hover:scale-125 dark:hover:text-white"
+          className="h-[20px] transition-all duration-300 hover:font-bold hover:text-purple-400 dark:hover:text-white max-lg:block max-lg:px-2 max-lg:py-1 max-lg:text-lg lg:text-base lg:hover:scale-125"
           onClick={onNavClicked}
         >
           <Link key={i} to={route.href}>
