--- client/src/constants/stateVariants.ts
@@ -1,1 +1,0 @@
-export const stateVariants = ['PRESENT_OPTIONAL', 'PRESENT_REQUIRED']
