+++ client/src/main.tsx
@@ -2,7 +2,7 @@ import ReactDOM from 'react-dom/client'
 import '@/styles/global.css'
 import { ThemeProvider } from './components/ThemeProvider.tsx'
 import { RouterProvider } from '@tanstack/react-router'
-import { router } from '@/routes'
+import { router } from '@/router/index.tsx'
 import { I18nextProvider } from 'react-i18next'
 import i18next from '@/translations'
 import { Provider } from 'react-redux'
